requirements.txt
pyTelegramBotAPI==4.21.0
python-dotenv==1.0.1
Requests==2.32.3
